3/2/EYAZIS/1/requirements.txt
PyQt5
nltk
psycopg2-binary
pymupdf
python-dotenv